Refactor: Break out components into modules
