+++ app/assets/scripts/components/documents/single-edit/local-store/index.js
@@ -1,101 +1,13 @@
 import React, { useState, useEffect, useCallback, useRef } from 'react';
 import T from 'prop-types';
 import { useFormikContext } from 'formik';
-import styled from 'styled-components';
+
 import { useParams } from 'react-router';
 import defaultsDeep from 'lodash.defaultsdeep';
 import { toast } from 'react-toastify';
-import { Button as BaseButton } from '@devseed-ui/button';
 
-export class LocalAtbdStorage {
-  constructor() {
-    this.expiry = 15 * 60000; // 15 minutes
-  }
-
-  getStorageKey(atbd, step) {
-    const { id, version } = atbd;
-    return `atbd/${id}/${version}/${step}`;
-  }
-
-  getAtbd(atbd, step) {
-    const key = this.getStorageKey(atbd, step);
-    const values = JSON.parse(localStorage.getItem(key));
-
-    if (values && values.created + this.expiry > Date.now()) {
-      // Return values if the store is not expired
-      return values;
-    }
-
-    // The store either doesn't exist or is expired,
-    // trying to remove the store
-    this.removeAtbd(atbd, step);
-    return undefined;
-  }
-
-  setAtbd(atbd, step, values) {
-    const key = this.getStorageKey(atbd, step);
-    localStorage.setItem(
-      key,
-      JSON.stringify({
-        ...values,
-        created: Date.now()
-      })
-    );
-  }
-
-  removeAtbd(atbd, step) {
-    const key = this.getStorageKey(atbd, step);
-    localStorage.removeItem(key);
-  }
-}
-
-const Button = styled(BaseButton)`
-  margin-top: 0.5rem;
-  margin-right: 0.5rem;
-`;
-
-function RecoverToast({ recoverData, clearData, closeToast }) {
-  const handleRecover = () => {
-    recoverData();
-    closeToast();
-  };
-
-  const handleClear = () => {
-    clearData();
-    closeToast();
-  };
-
-  return (
-    <div>
-      <p>
-        We have recovered data that you had inserted. Do you want to restore the
-        data?
-      </p>
-      <Button
-        variation='primary-raised-light'
-        size='small'
-        type='button'
-        onClick={handleRecover}
-      >
-        Recover
-      </Button>
-      <Button
-        variation='primary-raised-light'
-        size='small'
-        type='button'
-        onClick={handleClear}
-      >
-        Discard
-      </Button>
-    </div>
-  );
-}
-
-RecoverToast.propTypes = {
-  recoverData: T.func.isRequired,
-  clearData: T.func.isRequired,
-  closeToast: T.func.isRequired
-};
+import LocalAtbdStorage from './atbd-storage';
+import RecoverToast from './recover-toast';
 
 export function LocalStore({ atbd }) {
   const isInitialized = useRef();
